refactor(tui): extract host-result renderers to app/host_results.rs (S4 PR 2)
Second slice of the app.rs decomposition. All `append_*_host_result`
renderers, the `append_extension_host_result` dispatcher, plus the
related `append_host_command_error` / `append_legacy_cli_error` /
`append_memory_result` / `append_cli_result` / telegram-render helpers
move into a sibling submodule. No behaviour change; 84/84 TUI tests
pass.

Mechanics:
  - new file: crates/memphis-tui/src/app/host_results.rs (~1322 LoC)
  - `mod host_results;` added to app/mod.rs
  - 4 entry points pub(super) (callable from parent dispatch):
      append_host_command_error, append_extension_host_result,
      append_telegram_send_failure, append_telegram_send_cancelled
  - 3 renderers also pub(super) for non-host-RPC dispatch paths:
      append_legacy_cli_error, append_memory_result, append_cli_result
  - 28 per-command renderers stay private to host_results

LoC impact:
  - app/mod.rs: 5395 → 4100 (-1295)
  - app/host_results.rs: 0 → 1322 (+1322)
  - 27 unrelated lines diff is whitespace boundaries

Total app/ contents now: mod.rs 4100, commands.rs 671, host_results.rs 1323.
Still over the 1500-LoC ceiling on mod.rs; next S4 slices will pull out
formatting helpers + the test module.
